# project-notes/api-spec.yaml
openapi: 3.1.0
info:
  title: Project Allocation Management System (PAMS) API
  version: 1.12.0
  description: |
    REST API for the Project Allocation Management System.
    All endpoints require a Keycloak-issued OIDC Bearer token unless stated otherwise.
    Obtain tokens via the Keycloak Authorization Code flow at the configured realm (`/realms/pams`).
    Identity resolution uses the `empCode` JWT claim for Keycloak-to-PAMS identity bridging (sub claim verified against employees table with empCode fallback).
    Dates use ISO 8601 `YYYY-MM-DD` format. Timestamps use UTC (`YYYY-MM-DDTHH:MM:SSZ`).

    **Pagination (NFR-18/19):** All list endpoints support `page` (default 1) and `limit` (default 10, max 100).
    **Single-endpoint strategy (NFR-20):** Each resource has one list endpoint that also supports search and filter via query parameters.
    **Role-based response shaping (NFR-21):** Responses are filtered server-side based on the authenticated user's role.
    **Consolidated REST design:** Deactivation is handled via PUT (state change), allocation stop via PATCH, allocation remove via DELETE (soft-delete).
    **Billable flag & project status:** Projects carry a `billable` boolean (defaults to true for Client accounts) and a lifecycle `status` (Upcoming → Active → Completed). Both HR and PM (scoped to own projects) can modify status.
...
  contact:
    name: PAMS Engineering
  license:
    name: Internal Use Only
servers:
  - url: https://api.pams.internal/api/v1
...
security:
  - BearerAuth: []
...
paths:
...
  /employees:
    get:
      tags: [Employees]
      summary: List, search, and filter employees (HR, PM)
      description: |
        Unified endpoint for listing, searching, and filtering employees (NFR-20).
        Supports text search (name, empCode, skill), role filter, active status filter,
        bench-only filter, and date window for availability computation.
        When searching by skill, bench employees (0% allocated) are returned first.
      operationId: employees_list
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
            minLength: 2
          description: "Search query. Matches against full name (partial), empCode (prefix), or skill name (exact). Minimum 2 characters when provided."
          example: Android
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/EmployeeRole"
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: benchOnly
          in: query
          schema:
            type: boolean
          description: Return only employees with 0% allocation for the given date window.
        - name: windowFrom
          in: query
          schema:
            type: string
            format: date
          description: "Start of availability window (default: today)."
        - name: windowTo
          in: query
          schema:
            type: string
            format: date
          description: "End of availability window (default: today)."
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Paginated list of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponse_EmployeeSummary"
        "400":
          $ref: "#/components/responses/InvalidSortError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /employees/export:
    post:
      tags: [Employees]
      summary: Export employees as PDF or Excel (HR, PM)
      description: |
        Returns all matching employees (no pagination) as a downloadable PDF or Excel file.
        Accepts the same filter and sort query parameters as the list endpoint.
        An optional JSON body may specify a column name map to control which columns appear
        in the export and their display labels.
      operationId: employees_export
      parameters:
        - name: search
          in: query
          schema:
            type: string
            minLength: 2
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/EmployeeRole"
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: benchOnly
          in: query
          schema:
            type: boolean
        - name: windowFrom
          in: query
          schema:
            type: string
            format: date
        - name: windowTo
          in: query
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/ExportExtParam"
      requestBody:
        required: false
        description: |
          Optional column mapping. Keys are response DTO field names (case-insensitive),
          values are display labels for the export header. Only matched columns are included.
          If omitted or empty, all default columns are exported.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportColumnMap"
            example:
              empCode: "Employee Code"
              fullName: "Full Name"
              designation: "Title"
              isActive: "Active?"
      responses:
        "200":
          description: File download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="employees-2026-03-05.pdf"'
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      tags: [Employees]
      summary: Create a new employee (HR only)
      operationId: employees_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
            example:
              empCode: EMP042
              firstName: Jordan
              lastName: Lee
              email: jordan.lee@example.com
              designation: Senior Developer
              role: Staff
              reportsToEmpCode: EMP001
              skillIds:
                - c1d2e3f4-a5b6-7890-abcd-ef1234567890
                - d1e2f3a4-b5c6-7890-abcd-ef1234567890
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: empCode or email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://pams.internal/errors/ERR_EMPCODE_EXISTS"
                title: Conflict
                status: 409
                detail: "Employee code 'EMP042' is already in use."
        "422":
          description: Circular reporting chain detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: "https://pams.internal/errors/ERR_CIRCULAR_REPORTING"
                title: Unprocessable Entity
                status: 422
                detail: Setting this reportsTo would create a circular reporting chain.
  /employees/{empCode}:
    parameters:
      - $ref: "#/components/parameters/EmpCode"

    get:
      tags: [Employees]
      summary: Get employee by code (profile only)
      description: |
        Returns employee profile data (name, designation, role, skills, reporting chain).
        Does NOT include allocations or managed projects. Use `GET /allocations?empCode=...` to fetch allocations separately.
      operationId: employees_getByCode
      parameters:
        - name: windowFrom
          in: query
          schema:
            type: string
            format: date
          description: "Start of availability window (default: today)."
        - name: windowTo
          in: query
          schema:
            type: string
            format: date
          description: "End of availability window (default: today)."
      responses:
        "200":
          description: Employee detail with current allocation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Employees]
      summary: Update employee (HR only). empCode is immutable. Set isActive=false to deactivate.
      operationId: employees_update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetailResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Circular reporting chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /employees/me:
    get:
      tags: [Employees]
      summary: Get authenticated user's own profile (all roles)
      description: |
        Returns the employee profile for the currently authenticated user (profile only — no embedded allocations or managed projects).
        Use `GET /allocations?empCode=myEmpCode` for allocations and `GET /allocations?projectManagerEmpCode=myEmpCode` for managed-project allocations.
        Resolves the authenticated user via the `empCode` JWT claim and database lookup. See NFR-22 for identity resolution details.
      operationId: employees_me
      parameters:
        - name: windowFrom
          in: query
          schema:
            type: string
            format: date
          description: "Start of availability window (default: today)."
        - name: windowTo
          in: query
          schema:
            type: string
            format: date
          description: "End of availability window (default: today)."
      responses:
        "200":
          description: Authenticated user's full employee profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── EMPLOYEE SELF-MANAGE SKILLS (FR-023 – MVP2) ───────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC Bearer token issued by Keycloak (RS256 signed).
        Obtain via Keycloak Authorization Code flow: `{Authority}/protocol/openid-connect/auth`
        or Client Credentials flow for machine-to-machine.
        Default realm: `https://keycloak.internal/realms/pams`.

  # ── PARAMETERS ─────────────────────────────────────────────────────────────
  parameters:
...
    EmpCode:
      name: empCode
      in: path
      required: true
      schema:
        type: string
        maxLength: 50
      example: EMP042
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: "Number of items per page (default: 10, max: 100). Per NFR-18."
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: |
        Sort field. Use `fieldName` for ascending or `-fieldName` for descending.
        Each endpoint has a default sort if omitted. Only one sort field per request.
        Invalid field names return 400 with a ProblemDetails response containing
        `field` (the invalid field) and `allowedFields` (list of valid field names).
      example: "-fromDate"
    ExportExtParam:
      name: ext
      in: query
      required: true
      schema:
        type: string
        enum: [pdf, xls]
      description: "Export format: `pdf` for PDF or `xls` for Excel (.xlsx)."

  # ── RESPONSES ──────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_UNAUTHORIZED"
            title: Unauthorized
            status: 401
            detail: Authentication token is missing or invalid.
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_FORBIDDEN"
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action.
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_VALIDATION"
            title: Validation Failed
            status: 400
            errors:
              Percentage:
                - Percentage must be at least 25%.
                - Percentage must be a multiple of 5%.
...
    InvalidSortError:
      description: "Invalid sort field. Returns the field name and list of allowed sort fields."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InvalidSortProblem"
          example:
            type: "https://pams.internal/errors/ERR_INVALID_SORT"
            title: "Invalid Sort Parameter"
            status: 400
            detail: "Invalid sort field 'xyz'. Allowed fields: accountCode, accountName, accountType, isActive"
            instance: "/api/v1/accounts"
            field: "xyz"
            allowedFields: ["accountCode", "accountName", "accountType", "isActive"]

  # ── SCHEMAS ────────────────────────────────────────────────────────────────
  schemas:
...
    EmployeeRole:
      type: string
      enum: [HR, ProjectManager, Staff]
      description: Role of the employee within the system.
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string
...
    ExportColumnMap:
      type: object
      nullable: true
      additionalProperties:
        type: string
      description: |
        Column name mapping for export. Keys are response DTO field names (case-insensitive match),
        values are the display labels to use in the export header. Only matched columns are included.
        If null or empty, all default columns are exported.
      example:
        empCode: "Employee Code"
        fullName: "Name"

    # ── Pagination ─────────────────────────────────────────────────────────
...
    EmployeeDetailResponse:
      allOf:
        - $ref: "#/components/schemas/EmployeeSummary"
        - type: object
          properties:
            email:
              type: string
              format: email
            reportsToId:
              type: ["string", "null"]
              format: uuid
            reportsToEmpCode:
              type: ["string", "null"]
            reportsToName:
              type: ["string", "null"]
            skillDetails:
              type: array
              items:
                $ref: "#/components/schemas/SkillResponse"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    PagedResponse_EmployeeSummary:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSummary"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
    CreateEmployeeRequest:
      type: object
      required: [empCode, firstName, lastName, email, designation, role]
      properties:
        empCode:
          type: string
          maxLength: 50
        firstName:
          type: string
          maxLength: 100
        lastName:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        designation:
          type: string
          maxLength: 150
        role:
          $ref: "#/components/schemas/EmployeeRole"
        reportsToEmpCode:
          type: ["string", "null"]
          maxLength: 50
          description: Employee code of the reporting manager.
        skillIds:
          type: array
          items:
            type: string
            format: uuid
...